+++ test-projects/private-package/src/init.lua
@@ -1,0 +1,1 @@
+return "Private Package"