+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Wally Changelog
 
 ## Unreleased Changes
+* Added private field to package manifest ([#9])
+
+[#9]: https://github.com/UpliftGames/wally/issues/9
 
 ## 0.3.1 (2021-11-12)
 * Support for dev dependencies ([#63][63])